ref(discover) Move query results into a separate route
Since the addition of pagination and filtering saved queries we've been
running into more bugs that are caused by conflicting state between the
two modes of the page (query list & results). Moving the results to
a separate page lets us be more explicit about switching views and clear
out query string state.
