+++ src/sentry/static/sentry/app/routes.jsx
@@ -1179,6 +1179,13 @@ function routes() {
               component={errorHandler(LazyLoad)}
             />
             <Route
+              path="results/"
+              componentPromise={() =>
+                import(/* webpackChunkName: "DiscoverV2Results" */ 'app/views/eventsV2/results')
+              }
+              component={errorHandler(LazyLoad)}
+            />
+            <Route
               path=":eventSlug/"
               componentPromise={() =>
                 import(/* webpackChunkName: "DiscoverV2Details" */ 'app/views/eventsV2/eventDetails')
