+++ src/sentry/static/sentry/app/views/eventsV2/queryList.tsx
@@ -72,7 +72,7 @@ class QueryList extends React.Component<Props> {
 
   renderQueries() {
     const {pageLinks} = this.props;
-    const links = parseLinkHeader(pageLinks);
+    const links = parseLinkHeader(pageLinks || '');
     let cards: React.ReactNode[] = [];
 
     // If we're on the first page (no-previous page exists)
@@ -111,8 +111,9 @@ class QueryList extends React.Component<Props> {
         moment(eventView.start).format('MMM D, YYYY h:mm A') +
         ' - ' +
         moment(eventView.end).format('MMM D, YYYY h:mm A');
+
       const to = {
-        pathname: location.pathname,
+        pathname: `/organizations/${organization.slug}/eventsV2/results/`,
         query: {
           ...location.query,
           // remove any landing page cursor
@@ -167,7 +168,7 @@ class QueryList extends React.Component<Props> {
         ' - ' +
         moment(eventView.end).format('MMM D, YYYY h:mm A');
       const to = {
-        pathname: location.pathname,
+        pathname: `/organizations/${organization.slug}/eventsV2/results/`,
         query: {
           ...location.query,
           // remove any landing page cursor
