+++ tests/js/spec/views/eventsV2/index.spec.jsx
@@ -1,33 +1,9 @@
 import React from 'react';
 import {mountWithTheme} from 'sentry-test/enzyme';
 
-import {initializeOrg} from 'sentry-test/initializeOrg';
-
 import {DiscoverLanding} from 'app/views/eventsV2/landing';
 
-const FIELDS = [
-  {
-    field: 'title',
-    title: 'Custom Title',
-  },
-  {
-    field: 'timestamp',
-    title: 'Custom Time',
-  },
-  {
-    field: 'user',
-    title: 'Custom User',
-  },
-];
-
-const generateFields = () => {
-  return {
-    fieldnames: FIELDS.map(i => i.title),
-    field: FIELDS.map(i => i.field),
-  };
-};
-
-describe('EventsV2', function() {
+describe('EventsV2 > Landing', function() {
   const eventTitle = 'Oh no something bad';
   const features = ['events-v2'];
 
@@ -84,7 +60,7 @@ describe('EventsV2', function() {
     const wrapper = mountWithTheme(
       <DiscoverLanding
         organization={TestStubs.Organization({features})}
-        location={{query: {...generateFields()}}}
+        location={{query: {}}}
         router={{}}
       />,
       TestStubs.routerContext()
@@ -92,57 +68,5 @@ describe('EventsV2', function() {
 
     const content = wrapper.find('SentryDocumentTitle');
     expect(content.text()).toContain('You need at least one project to use this view');
-  });
-
-  it('pagination cursor should be cleared when making a search', function() {
-    const organization = TestStubs.Organization({
-      features,
-      projects: [TestStubs.Project()],
-    });
-
-    const initialData = initializeOrg({
-      organization,
-      router: {
-        location: {query: {...generateFields(), cursor: '0%3A50%3A0'}},
-      },
-    });
-
-    const wrapper = mountWithTheme(
-      <DiscoverLanding
-        organization={organization}
-        params={{orgId: organization.slug}}
-        location={initialData.router.location}
-        router={initialData.router}
-      />,
-      initialData.routerContext
-    );
-
-    // ensure cursor query string is initially present in the location
-
-    expect(initialData.router.location).toEqual({
-      query: {
-        ...generateFields(),
-        cursor: '0%3A50%3A0',
-      },
-    });
-
-    // perform a search
-
-    const search = wrapper.find('#smart-search-input').first();
-
-    search.simulate('change', {target: {value: 'geo:canada'}}).simulate('submit', {
-      preventDefault() {},
-    });
-
-    // cursor query string should be omitted from the query string
-
-    expect(initialData.router.push).toHaveBeenCalledWith({
-      pathname: undefined,
-      query: {
-        ...generateFields(),
-        query: 'geo:canada',
-        statsPeriod: '14d',
-      },
-    });
   });
 });
