+++ tests/js/spec/views/eventsV2/results.spec.jsx
@@ -1,0 +1,125 @@
+import React from 'react';
+import {mountWithTheme} from 'sentry-test/enzyme';
+
+import {initializeOrg} from 'sentry-test/initializeOrg';
+
+import Results from 'app/views/eventsV2/results';
+
+const FIELDS = [
+  {
+    field: 'title',
+    title: 'Custom Title',
+  },
+  {
+    field: 'timestamp',
+    title: 'Custom Time',
+  },
+  {
+    field: 'user',
+    title: 'Custom User',
+  },
+];
+
+const generateFields = () => {
+  return {
+    fieldnames: FIELDS.map(i => i.title),
+    field: FIELDS.map(i => i.field),
+  };
+};
+
+describe('EventsV2 > Results', function() {
+  const eventTitle = 'Oh no something bad';
+  const features = ['events-v2'];
+
+  beforeEach(function() {
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/eventsv2/',
+      body: {
+        meta: {
+          id: 'string',
+          title: 'string',
+          'project.name': 'string',
+          timestamp: 'date',
+          'user.id': 'string',
+        },
+        data: [
+          {
+            id: 'deadbeef',
+            'user.id': 'alberto leal',
+            title: eventTitle,
+            'project.name': 'project-slug',
+            timestamp: '2019-05-23T22:12:48+00:00',
+          },
+        ],
+      },
+    });
+    MockApiClient.addMockResponse({
+      url: '/organizations/org-slug/events/project-slug:deadbeef/',
+      method: 'GET',
+      body: {
+        id: '1234',
+        size: 1200,
+        eventID: 'deadbeef',
+        title: 'Oh no something bad',
+        message: 'It was not good',
+        dateCreated: '2019-05-23T22:12:48+00:00',
+        entries: [
+          {
+            type: 'message',
+            message: 'bad stuff',
+            data: {},
+          },
+        ],
+        tags: [{key: 'browser', value: 'Firefox'}],
+      },
+    });
+  });
+
+  it('pagination cursor should be cleared when making a search', function() {
+    const organization = TestStubs.Organization({
+      features,
+      projects: [TestStubs.Project()],
+    });
+
+    const initialData = initializeOrg({
+      organization,
+      router: {
+        location: {query: {...generateFields(), cursor: '0%3A50%3A0'}},
+      },
+    });
+
+    const wrapper = mountWithTheme(
+      <Results
+        organization={organization}
+        location={initialData.router.location}
+        router={initialData.router}
+      />,
+      initialData.routerContext
+    );
+
+    // ensure cursor query string is initially present in the location
+    expect(initialData.router.location).toEqual({
+      query: {
+        ...generateFields(),
+        cursor: '0%3A50%3A0',
+      },
+    });
+
+    // perform a search
+    const search = wrapper.find('#smart-search-input').first();
+
+    search.simulate('change', {target: {value: 'geo:canada'}}).simulate('submit', {
+      preventDefault() {},
+    });
+
+    // cursor query string should be omitted from the query string
+    expect(initialData.router.push).toHaveBeenCalledWith({
+      pathname: undefined,
+      query: {
+        ...generateFields(),
+        query: 'geo:canada',
+        statsPeriod: '14d',
+      },
+    });
+  });
+});
